test: provide test setup variant 1 to create our app in test environment
Not the best way, at least not the "Nest Way". This will be updated in the next commit. This exist just to know that this setup is available. Stephen Grider himself said that he would use this instead of using the Nest setup. But since it's a course about Nest, he of course have the responsibility to teach you the right way to set things up.